src/ingest.py
import os
from PyPDF2 import PdfReader

def load_documents(folder_path):
    texts = []
    for filename in os.listdir(folder_path):
        if filename.endswith(".pdf"):
            reader = PdfReader(os.path.join(folder_path, filename))
            text = ""
            for page in reader.pages:
                text += page.extract_text()
            texts.append(text)
    return texts
