requirements.txt
faiss-cpu
langchain
sentence-transformers
streamlit
PyMuPDF
openai
chromadb
